restartCampAPI/requirements.txt
asgiref==3.3.4
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==3.10.0
numpy==1.19.5
pandas==1.1.5
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
XlsxWriter==1.3.8
zipp==3.4.1
django-cors-headers==3.7.0
